Book/backend/requirements.txt
flask
flask-cors
pyfiglet
openai
python-docx
langchain
